RUB-306 isolate suite registry tests
